% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryUtils.r
\name{logmod}
\alias{logmod}
\title{Make Trinary maps from previous workflow runs}
\usage{
logmod(x)
}
\value{
a data.frame
}
\description{
Read in workflow info, make trinary maps and plot vs thresholds
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
