% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryMap}
\alias{trinaryMap}
\title{Make trinary maps}
\usage{
trinaryMap(model, threshLo, threshHi, threshYouden, modelNames, species,
  rasterOutputDir = NULL, ...)
}
\value{
a data.frame
}
\description{
Use previously calculated thresholds to make trinary maps
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
