% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryWorkflow.r
\name{trinaryMapWorkflow}
\alias{trinaryMapWorkflow}
\title{Apply trinary map tools to a set of continuous range maps}
\usage{
trinaryMapWorkflow(
  pres,
  background,
  model,
  modelNames,
  species,
  rasterOutputDir = NULL,
  mapPlotDir = NULL,
  ROCPlotDir = NULL,
  shapesToPlot = NULL,
  openFig = T,
  NATo0 = TRUE
)
}
\arguments{
\item{expertRaster.f=NULL}{}
}
\value{
a data.frame
}
\description{
Run all aspects of modeling and plotting for a suite of previously run models
}
\details{
See Examples.
}
\note{
This is a convenience wrapper around each component of trinary map building.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
