% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryRangeSize}
\alias{trinaryRangeSize}
\title{Calculate upper and lower limits of range size}
\usage{
trinaryRangeSize(trinary.rasters, youden.binary.tmp.raster = NULL)
}
\arguments{
\item{youden.binary.tmp.raster}{}
}
\value{
a data.frame with the lower, upper, and optimal (Youden threshold) range size
}
\description{
Size limits based on trinary thresholds
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
