% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryPlots.r
\name{trinaryROCPlot}
\alias{trinaryROCPlot}
\title{Make Trinary maps from previous workflow runs}
\usage{
trinaryROCPlot(
  ROCPlotFile = NULL,
  trinaryPlotThings,
  trinaryDF,
  openFig = TRUE
)
}
\value{
a data.frame
}
\description{
Read in workflow info, make trinary maps and plot vs thresholds
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
