% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryROCRoots}
\alias{trinaryROCRoots}
\title{Find upper/lower thresholds from ROC curves}
\usage{
trinaryROCRoots(ins, max.sens = 0.95)
}
\arguments{
\item{max.sens=0.95}{}
}
\value{
a data.frame
}
\description{
Fit a smoothed ROC curve, find bounds for threshold and report partial AUC statistics
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
