% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryMap}
\alias{trinaryMap}
\title{Make trinary maps}
\usage{
trinaryMap(rModel, thr.roc.lo, thr.roc.hi, rasterOutputPath = NULL, ...)
}
\arguments{
\item{rModel}{spatRaster representing continuous model predictions}

\item{thr.roc.lo}{lower threshold value; typically determined from the 
output of `trinaryROCRoots()`}

\item{thr.roc.hi}{upper threshold value; typically determined from the 
output of `trinaryROCRoots()`}

\item{rasterOutputPath}{optional file name to write out a raster.}

\item{...}{optional arguments to pass to `terra::writeRaster`}
}
\value{
a data.frame
}
\description{
Use previously calculated thresholds to make trinary maps
}
\author{
Cory Merow <cory.merow@gmail.com>
}
