% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryPlots.r
\name{trinaryMapPlot}
\alias{trinaryMapPlot}
\title{Make Trinary maps from previous workflow runs}
\usage{
trinaryMapPlot(
  trinary.rasters,
  plotFile = NULL,
  pres = NULL,
  species,
  shapesToPlot = NULL,
  openFig = T
)
}
\value{
a data.frame
}
\description{
Read in workflow info, make trinary maps and plot vs thresholds
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
