% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryWorkflow.r
\name{trinaryMapWorkflow}
\alias{trinaryMapWorkflow}
\title{Apply trinary map tools to a set of continuous range maps}
\usage{
trinaryMapWorkflow(pres, background, model, modelNames, species,
  rasterOutputDir = NULL, mapPlotDir = NULL, ROCPlotDir = NULL,
  expertShpPath = NULL, expertRasterPath = NULL, shapesToPlot = NULL,
  openFig = T, doMapPlot = TRUE, doROCPlot = TRUE)
}
\arguments{
\item{expertRaster.f=NULL}{}
}
\value{
a data.frame
}
\description{
Run all aspects of modeling and plotting for a suite of previously run models
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
