% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryWorkflow.r
\name{trinaryMapWorkflow}
\alias{trinaryMapWorkflow}
\title{Generate trinary maps from species presence, background and model}
\usage{
trinaryMapWorkflow(pres, background, rModel, NATo0 = TRUE)
}
\arguments{
\item{pres}{A data.frame object containing the species presence data}

\item{background}{A data.frame object containing the background data}

\item{rModel}{A spatRaster object containing the model data}

\item{NATo0}{A logical indicating whether to turn NA values into zeros in the presence and background data}
}
\value{
A list containing the trinary thresholds and trinary rasters
}
\description{
This function takes in the species presence and background data, along with a model, and generates trinary maps. It also calculates the range size and plots the maps and ROC curves.
}
