% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryRangeSize}
\alias{trinaryRangeSize}
\title{Calculate upper and lower limits of range size}
\usage{
trinaryRangeSize(trinaryRasters)
}
\arguments{
\item{trinaryRasters}{a spatRaster describing a trinary map. It is assumed that 
values of 0 are absent, values of 1 represent the upper ROC bound, and values 
of 2 represent the lower ROC bound (e.g., as determined by `trinaryROCRoots`())}
}
\value{
a data.frame with the lower and upper ROC bounds.
}
\description{
Size limits based on trinary thresholds
}
\author{
Cory Merow <cory.merow@gmail.com>, Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>
}
