% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryRangeSize}
\alias{trinaryRangeSize}
\title{Calculate upper and lower limits of range size}
\usage{
trinaryRangeSize(trinaryRasters, otherBinaryRaster = NULL)
}
\arguments{
\item{trinaryRasters}{a spatRaster describing a trinary map. It is assumed that 
values of 0 are absent, values of 1 represent the lower bound, and values 
of 2 represent the upper bound (e.g., as determined by `trinaryROCRoots`())}

\item{otherBinaryRaster}{optional additional binary raster for which range 
size is desired. This can be useful, e.g., if you have an additional 
intermediate threshold of interest (e.g., the max (sensitivity +specificity)).}
}
\value{
a data.frame with the lower, upper, and optimal (Youden threshold) range size
}
\description{
Size limits based on trinary thresholds
}
\author{
Cory Merow <cory.merow@gmail.com>
}
