% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryPlots.r
\name{trinaryROCPlot}
\alias{trinaryROCPlot}
\title{Plot trinary ROC curve and save the output as pdf file.}
\usage{
trinaryROCPlot(trinaryPlotThings, trinaryDF, plotFile = NULL, openFig = TRUE)
}
\arguments{
\item{trinaryPlotThings}{Object of trinary ROC curve, contains xx, y, xout, and y. as its attributes.}

\item{trinaryDF}{Data frame with trinary ROC curve information, including youden.thresh.roc.x, hi.thresh.roc.x, lo.thresh.roc.x, and trinary.pauc.}

\item{plotFile}{Path to save the output as pdf file. If NULL, the plot will not be saved.}

\item{openFig}{Logical, if TRUE, the plot will be shown in a new window.}
}
\value{
None
}
\description{
Plot trinary ROC curve and save the output as pdf file.
}
