% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryROCRoots}
\alias{trinaryROCRoots}
\title{Find upper/lower thresholds from ROC curves}
\usage{
trinaryROCRoots(ins, max.sens = 0.95, smoothMethod = "binormal", ...)
}
\arguments{
\item{max.sens}{value of sensitivity to use if no upper limit is found based on derviatives. default is 0.95}

\item{smoothMethod}{default is 'binormal'}

\item{...}{options to be passed to `pROC::smooth`}
}
\value{
a data.frame
}
\description{
Fit a smoothed ROC curve, find bounds for threshold and report partial AUC statistics
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
