% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinaryModeling.r
\name{trinaryROCRoots}
\alias{trinaryROCRoots}
\title{Find upper/lower thresholds from ROC curves}
\usage{
trinaryROCRoots(
  ins,
  max.sens = 0.95,
  smoothMethod = "binormal",
  sdMultiplier = 2,
  maxTPQuantile = 0.3,
  ...
)
}
\arguments{
\item{ins}{A data frame where the first column contains the the observed 
binary outcomes (e.g., 0 for absence and 1 for presence) and the second column 
contains predictor values (e.g., predicted probabilities/suitability). This 
input is used to calculate the ROC curve, derive thresholds, and assess model 
performance.}

\item{max.sens}{value of sensitivity to use if no upper limit is found based 
on derivatives. default is 0.95}

\item{smoothMethod}{default is 'binormal'}

\item{sdMultiplier}{A numeric value that controls the adjustment of the lower 
threshold for presence predictions. It multiplies the standard deviation of 
the predicted values at presence locations to define this lower threshold. 
If this threshold results in a negative value, it is adjusted to zero 
by default. Default is `2`.}

\item{maxTPQuantile}{A numeric value representing the quantile of the predicted 
values at presence locations used to determine the upper threshold for 
presence predictions. For example, a value of `0.3` means the function 
uses the 30th percentile of the predicted presence values as the upper 
threshold, ensuring that the model maintains at least 70% 
sensitivity. Default is `0.3`.}

\item{...}{options to be passed to `pROC::smooth`}
}
\value{
a data.frame
}
\description{
Fit a smoothed ROC curve, find bounds for threshold and report 
 partial AUC statistics
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
